be/requirements.txt
fastapi
uvicorn
litellm
Faker
fpdf2
chromadb
PyPDF2
